made config directory with files for config class
closes #12
