+++ ai_presenter/config/ai.py
@@ -1,0 +1,6 @@
+from ai_presenter.config.chatgpt import ChatGPTConfig
+
+
+class AIConfig:
+    def __init__(self, data):
+        self.chatgptconfig = ChatGPTConfig(data['chatgpt_config'])
