+++ ai_presenter/config/chatgpt.py
@@ -1,0 +1,3 @@
+class ChatGPTConfig:
+    def __init__(self, data):
+        self.style = data['style']
