+++ ai_presenter/config/config.py
@@ -1,0 +1,6 @@
+from ai_presenter.config.ai import AIConfig
+
+
+class Config:
+    def __init__(self, data):
+        self.ai_config = AIConfig(data['ai_config'])
